✨ Metrics reporting (#10)
* ✨ Metrics reporting

* 💚 Install merged jars

+++ tools/library/lexakai-library-functions.sh
@@ -60,36 +60,48 @@ showVersion() {
 
 clean_cache() {
 
-    cache=$1
+    if [[ "$ALLOW_CLEANING" == "true" ]]; then
 
-    if [ -d "$cache" ]; then
-        read -p "┋ Remove ALL cached files in $cache (y/n)? " -n 1 -r
-        echo "┋ "
-        if [[ $REPLY =~ ^[Yy]$ ]]; then
-            rm -rf $cache
+        cache=$1
+
+        if [ -d "$cache" ]; then
+            read -p "┋ Remove ALL cached files in $cache (y/n)? " -n 1 -r
+            echo "┋ "
+            if [[ $REPLY =~ ^[Yy]$ ]]; then
+                rm -rf $cache
+            fi
         fi
+
     fi
 }
 
 clean_maven_repository() {
 
-    project_home=$1
-    name=$(basename -- "$project_home")
+    if [[ "$ALLOW_CLEANING" == "true" ]]; then
 
-    if yes_no "Remove all $name artifacts from ~/.m2"; then
+        project_home=$1
+        name=$(basename -- "$project_home")
 
-        rm -rf "~/.m2/repository/com/telenav/$name"
+        if yes_no "Remove all $name artifacts from ~/.m2"; then
+
+            rm -rf "~/.m2/repository/com/telenav/$name"
+
+        fi
 
     fi
 }
 
 remove_maven_repository() {
 
-    if [ -d "$HOME/.m2/repository" ]; then
+    if [[ "$ALLOW_CLEANING" == "true" ]]; then
 
-        if yes_no "Remove ALL artifacts in ~/.m2/repository"; then
+        if [ -d "$HOME/.m2/repository" ]; then
 
-            rm -rf ~/.m2/repository
+            if yes_no "Remove ALL artifacts in ~/.m2/repository"; then
+
+                rm -rf ~/.m2/repository
+
+            fi
 
         fi
 
@@ -98,11 +110,15 @@ remove_maven_repository() {
 
 clean_temporary_files() {
 
-    project_home=$1
+    if [[ "$ALLOW_CLEANING" == "true" ]]; then
 
-    if yes_no "Remove temporary files (.DS_Store, .metadata, .classpath, .project, *.hprof, *~) from $project_home tree"; then
+        project_home=$1
 
-        find $project_home \( -name \.DS_Store -o -name \.metadata -o -name \.classpath -o -name \.project -o -name \*\.hprof -o -name \*~ \) | xargs rm
+        if yes_no "Remove temporary files (.DS_Store, .metadata, .classpath, .project, *.hprof, *~) from $project_home tree"; then
+
+            find $project_home \( -name \.DS_Store -o -name \.metadata -o -name \.classpath -o -name \.project -o -name \*\.hprof -o -name \*~ \) | xargs rm
+
+        fi
 
     fi
 }
